Some updates to the Content component, specifically adding support for View of types json and xml. Some more work on the Navigation model. A LOT of fixes in the CurrentPage logic, including adding support for content types and better URL translation support.

+++ components/Content/templates/pages/content/index.tpl
@@ -16,7 +16,7 @@
 				<ul class="controls">
 					<li class="view">{a href=$page->getLink('Page')->get('rewriteurl')}View{/a}</li>
 					<li class="edit">{a href="/Content/Edit/`$page->get('id')`"}Edit{/a}</li>
-					<li class="delete">{a href="/Content/Delete/`$page->get('id')`"}Delete{/a}</li>
+					<li class="delete">{a href="/Content/Delete/`$page->get('id')`" confirm="Are you sure you want to delete `$page->get('title')`?"}Delete{/a}</li>
 				</ul>
 			</td>
 		</tr>
